refactor(context): move context into sub directory after and before
