+++ src/DatasourceCustomizer/Decorators/Hook/Context/After/HookAfterCreateContext.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace ForestAdmin\AgentPHP\DatasourceCustomizer\Decorators\Hook\Context;
+namespace ForestAdmin\AgentPHP\DatasourceCustomizer\Decorators\Hook\Context\After;
 
+use ForestAdmin\AgentPHP\DatasourceCustomizer\Decorators\Hook\Context\Before\HookBeforeCreateContext;
 use ForestAdmin\AgentPHP\DatasourceToolkit\Components\Contracts\CollectionContract;
 
 class HookAfterCreateContext extends HookBeforeCreateContext
