+++ src/DatasourceCustomizer/Decorators/Hook/Context/After/HookAfterListContext.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace ForestAdmin\AgentPHP\DatasourceCustomizer\Decorators\Hook\Context;
+namespace ForestAdmin\AgentPHP\DatasourceCustomizer\Decorators\Hook\Context\After;
 
+use ForestAdmin\AgentPHP\DatasourceCustomizer\Decorators\Hook\Context\Before\HookBeforeListContext;
 use ForestAdmin\AgentPHP\DatasourceToolkit\Components\Contracts\CollectionContract;
 use ForestAdmin\AgentPHP\DatasourceToolkit\Components\Query\Filters\PaginatedFilter;
 use ForestAdmin\AgentPHP\DatasourceToolkit\Components\Query\Projection\Projection;
