+++ src/DatasourceCustomizer/Decorators/Hook/Context/After/HookAfterUpdateContext.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace ForestAdmin\AgentPHP\DatasourceCustomizer\Decorators\Hook\Context\After;
+
+use ForestAdmin\AgentPHP\DatasourceCustomizer\Decorators\Hook\Context\Before\HookBeforeUpdateContext;
+
+class HookAfterUpdateContext extends HookBeforeUpdateContext
+{
+}
