+++ src/DatasourceCustomizer/Decorators/Hook/Context/Before/HookBeforeCreateContext.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace ForestAdmin\AgentPHP\DatasourceCustomizer\Decorators\Hook\Context;
+namespace ForestAdmin\AgentPHP\DatasourceCustomizer\Decorators\Hook\Context\Before;
 
+use ForestAdmin\AgentPHP\DatasourceCustomizer\Decorators\Hook\Context\HookContext;
 use ForestAdmin\AgentPHP\DatasourceToolkit\Components\Contracts\CollectionContract;
 
 class HookBeforeCreateContext extends HookContext
