--- src/DatasourceCustomizer/Decorators/Hook/Context/HookAfterDeleteContext.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace ForestAdmin\AgentPHP\DatasourceCustomizer\Decorators\Hook\Context;
-
-class HookAfterDeleteContext extends HookBeforeDeleteContext
-{
-}
