--- src/DatasourceCustomizer/Decorators/Hook/Context/HookAfterUpdateContext.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace ForestAdmin\AgentPHP\DatasourceCustomizer\Decorators\Hook\Context;
-
-class HookAfterUpdateContext extends HookBeforeUpdateContext
-{
-}
